[memprof] Move YAML traits to MemProf.h (NFC)
This patch moves the MemProf YAML traits to MemProf.h so that the YAML
writer can access them from outside MemProfReader.cpp in the future.
